requirements.txt
fastapi>=0.109.0,<1.0.0
uvicorn>=0.27.0,<1.0.0
python-dotenv>=1.0.0,<2.0.0
pydantic>=2.0.0,<3.0.0
openai>=1.0.0,<2.0.0
langchain-openai>=0.1.0,<2.0.0
langchain==1.2.15
langgraph==1.1.6
